LayoutTests/media/video-timeupdate-during-playback-expected.txt

Test 'timeupdate' events are posted while playing but not while paused.

RUN(video.play())

EVENT(play)
EVENT(waiting)
EVENT(loadstart)
EVENT(durationchange)
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
EVENT(playing)

RUN(video.pause())
EVENT(pause)

END OF TEST
